+++ .eslintrc.js
@@ -55,6 +55,10 @@ module.exports = {
       minLength: 4,
     }],
 
+    // Disables double quote error when using single quotes within string for readability
+    // https://eslint.org/docs/rules/quotes#avoidescape
+    'quotes': ['error', 'single', { 'avoidEscape': true }],
+
     // Typescript rules
     // Extends recommended rules here: https://www.npmjs.com/package/@typescript-eslint/eslint-plugin
     '@typescript-eslint/naming-convention': [ 'error',
@@ -179,6 +183,7 @@ module.exports = {
       rules: {
         '@typescript-eslint/unbound-method': 'off',
         'jest/unbound-method': 'error',
+        '@typescript-eslint/no-explicit-any': 'off',
       }
     }
   ],
@@ -213,6 +218,7 @@ module.exports = {
     '/packages/amplify-category-function/lib',
     '/packages/amplify-category-geo/lib',
     '/packages/amplify-category-storage/lib',
+    '/packages/amplify-cli-npm/lib',
     '/packages/amplify-cli-core/lib',
     '/packages/amplify-cli/lib',
     '/packages/amplify-cli-logger/lib',