+++ packages/amplify-nodejs-function-template-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.10](https://github.com/aws-amplify/amplify-cli/compare/amplify-nodejs-function-template-provider@2.3.9...amplify-nodejs-function-template-provider@2.3.10) (2022-04-07)
+
+**Note:** Version bump only for package amplify-nodejs-function-template-provider
+
+
+
+
+
 ## [2.3.9](https://github.com/aws-amplify/amplify-cli/compare/amplify-nodejs-function-template-provider@2.3.8...amplify-nodejs-function-template-provider@2.3.9) (2022-03-23)
 
 **Note:** Version bump only for package amplify-nodejs-function-template-provider